requirements.txt
numpy
librosa
soundfile
faster-whisper
pyopenssl
torchaudio
rapidfuzz
grpcio

